# be/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse:
    properties:
      error_code:
        type: string
      error_message:
        type: string
    type: object
...
  github_com_andreypavlenko_jobber_internal_platform_http.PaginatedResponse:
    properties:
      items: {}
      pagination:
        $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.PaginationMeta'
    type: object
  github_com_andreypavlenko_jobber_internal_platform_http.PaginationMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_andreypavlenko_jobber_modules_jobs_model.CreateJobRequest:
    properties:
      applied_at:
        type: string
      company_id:
        type: string
      description:
        type: string
      notes:
        type: string
      resume_builder_id:
        type: string
      resume_id:
        type: string
      source:
        type: string
      status:
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
      url:
        type: string
    required:
    - title
    type: object
...
  github_com_andreypavlenko_jobber_modules_jobs_model.JobDTO:
    properties:
      applied_at:
        type: string
      company_id:
        type: string
      company_name:
        type: string
      created_at:
        type: string
      current_stage_id:
        type: string
      current_stage_name:
        type: string
      description:
        type: string
      id:
        type: string
      is_favorite:
        type: boolean
      job_comments:
        items:
          $ref: '#/definitions/github_com_andreypavlenko_jobber_modules_comments_model.CommentDTO'
        type: array
      last_activity_at:
        type: string
      notes:
        type: string
      resume:
        $ref: '#/definitions/github_com_andreypavlenko_jobber_modules_jobs_model.ResumeNestedDTO'
      source:
        type: string
      stage_comments:
        items:
          $ref: '#/definitions/github_com_andreypavlenko_jobber_modules_comments_model.CommentDTO'
        type: array
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  github_com_andreypavlenko_jobber_modules_jobs_model.MoveJobRequest:
    properties:
      target:
        $ref: '#/definitions/github_com_andreypavlenko_jobber_modules_jobs_model.MoveTarget'
    required:
    - target
    type: object
...
  github_com_andreypavlenko_jobber_modules_jobs_model.UpdateJobRequest:
    properties:
      applied_at:
        type: string
      company_id:
        type: string
      description:
        type: string
      notes:
        type: string
      resume_builder_id:
        type: string
      resume_id:
        type: string
      source:
        type: string
      status:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@jobber.example.com
    name: API Support
  description: Job Application Tracking Platform API - A modular monolith backend
    for managing job applications, companies, resumes, and application stages.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Jobber API
  version: "1.0"
paths:
...
  /jobs:
    get:
      description: Get a paginated list of job cards for the authenticated user with
        filtering and sorting
      parameters:
      - description: 'Number of items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of items to skip (default: 0)'
        in: query
        name: offset
        type: integer
      - description: 'Filter by status: saved, applied, on_hold, offer, rejected,
          archived, all. Empty and the legacy value ''active'' mean everything except
          archived.'
        in: query
        name: status
        type: string
      - description: 'Sort format: field:order (e.g., last_activity:desc, created_at:desc,
          title:asc, company_name:asc, status:asc, applied_at:desc)'
        in: query
        name: sort
        type: string
      - description: Case-insensitive search matched against job title and company
          name (max 100 chars)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.PaginatedResponse'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/github_com_andreypavlenko_jobber_modules_jobs_model.JobDTO'
                  type: array
              type: object
        "400":
          description: Invalid pagination parameters
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List jobs
      tags:
      - jobs
    post:
      consumes:
      - application/json
      description: Create a new job card for the authenticated user. Without status/resume
        fields a "saved" (wishlist) card is created.
      parameters:
      - description: Job details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_andreypavlenko_jobber_modules_jobs_model.CreateJobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_modules_jobs_model.JobDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new job
      tags:
      - jobs
  /jobs/{id}:
    delete:
      description: Delete a specific job card by ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "404":
          description: Job not found
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a job
      tags:
      - jobs
    get:
      description: Get details of a specific job by ID, including pipeline state and
        comments
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_modules_jobs_model.JobDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "404":
          description: Job not found
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a job
      tags:
      - jobs
    patch:
      consumes:
      - application/json
      description: Update details or pipeline status of a specific job. Moving out
        of "saved" stamps applied_at; moving back clears it.
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated job details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_andreypavlenko_jobber_modules_jobs_model.UpdateJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_modules_jobs_model.JobDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "404":
          description: Job not found
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a job
      tags:
      - jobs
...
  /jobs/{id}/favorite:
    post:
      description: Toggle the favorite status of a specific job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "404":
          description: Job not found
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle job favorite status
      tags:
      - jobs
  /jobs/{id}/move:
    post:
      consumes:
      - application/json
      description: 'Atomically reposition a card: a stage target sets the stage and
        derives the status from the stage''s phase; a phase target completes the current
        stage and sets the derived status'
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Move target
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_andreypavlenko_jobber_modules_jobs_model.MoveJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_modules_jobs_model.JobDTO'
        "400":
          description: Invalid move target or phase
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "404":
          description: Job or stage template not found
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_andreypavlenko_jobber_internal_platform_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Move a job on the unified board
      tags:
      - jobs
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...